modules/gvi_segmentation/requirements.txt
streamlit
matplotlib
torch
pillow
transformers
numpy